requirements.txt
Django>=4.2
djangorestframework
requests
psycopg2-binary
gunicorn
